Keep @aws-sdk out of the static graph behind files-sdk/r2
